+++ Modules/SharePointDsc/DSCResources/MSFT_SPManagedMetadataServiceApp/MSFT_SPManagedMetaDataServiceApp.psm1
@@ -5,6 +5,7 @@ function Get-TargetResource
     param
     (
         [parameter(Mandatory = $true)]  [System.String] $Name,
+        [parameter(Mandatory = $false)] [System.String] $ProxyName,
         [parameter(Mandatory = $true)]  [System.String] $ApplicationPool,
         [parameter(Mandatory = $false)] [System.String] $DatabaseServer,
         [parameter(Mandatory = $false)] [System.String] $DatabaseName,
@@ -26,7 +27,7 @@ function Get-TargetResource
         } 
         if ($null -eq $serviceApps) { 
             return $nullReturn 
-        }
+      }
         $serviceApp = $serviceApps | Where-Object { $_.TypeName -eq "Managed Metadata Service" }
 
         if ($null -eq $serviceApp)
@@ -35,8 +36,15 @@ function Get-TargetResource
         }
         else
         {
+        $serviceAppProxies = Get-SPServiceApplicationProxy -ErrorAction SilentlyContinue
+        if ($null -ne $serviceAppProxies)
+        {
+            $serviceAppProxy = $serviceAppProxies | Where-Object { $_.GetType().Name -eq 'MetadataWebServiceApplicationProxy' }
+            if ($null -ne $serviceAppProxy) { $proxyName = $serviceAppProxy.Name}
+        }
             return @{
                 Name            = $serviceApp.DisplayName
+                ProxyName       = $proxyName
                 Ensure          = "Present"
                 ApplicationPool = $serviceApp.ApplicationPool.Name
                 DatabaseName    = $serviceApp.Database.Name
@@ -55,6 +63,7 @@ function Set-TargetResource
     param
     (
         [parameter(Mandatory = $true)]  [System.String] $Name,
+        [parameter(Mandatory = $false)] [System.String] $ProxyName,
         [parameter(Mandatory = $true)]  [System.String] $ApplicationPool,
         [parameter(Mandatory = $false)] [System.String] $DatabaseServer,
         [parameter(Mandatory = $false)] [System.String] $DatabaseName,
@@ -76,11 +85,13 @@ function Set-TargetResource
                 $params.Add("HubUri", $params.ContentTypeHubUrl)
                 $params.Remove("ContentTypeHubUrl")
             }
+            if ($params.ContainsKey("ProxyName")) { $pName = $params.ProxyName ; $params.Remove("ProxyName") | Out-Null }
+            if ($null -eq $pName) {$pName = "$($params.Name) Proxy"}
 
             $app = New-SPMetadataServiceApplication @params 
             if ($null -ne $app)
             {
-                New-SPMetadataServiceApplicationProxy -Name ($params.Name + " Proxy") `
+                New-SPMetadataServiceApplicationProxy -Name $pName `
                                                       -ServiceApplication $app `
                                                       -DefaultProxyGroup `
                                                       -ContentTypePushdownEnabled `
@@ -124,6 +135,7 @@ function Test-TargetResource
     param
     (
         [parameter(Mandatory = $true)]  [System.String] $Name,
+        [parameter(Mandatory = $false)] [System.String] $ProxyName,
         [parameter(Mandatory = $true)]  [System.String] $ApplicationPool,
         [parameter(Mandatory = $false)] [System.String] $DatabaseServer,
         [parameter(Mandatory = $false)] [System.String] $DatabaseName,