+++ Modules/SharePointDsc/DSCResources/MSFT_SPSecureStoreServiceApp/MSFT_SPSecureStoreServiceApp.psm1
@@ -5,6 +5,7 @@ function Get-TargetResource
     param
     (
         [parameter(Mandatory = $true)]  [System.String]  $Name,
+        [parameter(Mandatory = $false)] [System.String]  $ProxyName,
         [parameter(Mandatory = $true)]  [System.String]  $ApplicationPool,
         [parameter(Mandatory = $true)]  [System.Boolean] $AuditingEnabled,
         [parameter(Mandatory = $false)] [System.UInt32]  $AuditlogMaxSize,
@@ -60,6 +61,7 @@ function Set-TargetResource
     param
     (
         [parameter(Mandatory = $true)]  [System.String]  $Name,
+        [parameter(Mandatory = $false)] [System.String]  $ProxyName,
         [parameter(Mandatory = $true)]  [System.String]  $ApplicationPool,
         [parameter(Mandatory = $true)]  [System.Boolean] $AuditingEnabled,
         [parameter(Mandatory = $false)] [System.UInt32]  $AuditlogMaxSize,
@@ -101,7 +103,10 @@ function Set-TargetResource
                 $params.Remove("DatabaseAuthenticationType")
             }
 
-            New-SPSecureStoreServiceApplication @params | New-SPSecureStoreServiceApplicationProxy -Name "$($params.Name) Proxy"
+            if ($params.ContainsKey("ProxyName")) { $pName = $params.ProxyName ; $params.Remove("ProxyName") | Out-Null }
+            if ($pName -eq $Null) {$pName = "$($params.Name) Proxy"}
+            
+            New-SPSecureStoreServiceApplication @params | New-SPSecureStoreServiceApplicationProxy -Name $pName
         }
     } 
     
@@ -138,6 +143,7 @@ function Test-TargetResource
     param
     (
         [parameter(Mandatory = $true)]  [System.String]  $Name,
+        [parameter(Mandatory = $false)] [System.String]  $ProxyName,
         [parameter(Mandatory = $true)]  [System.String]  $ApplicationPool,
         [parameter(Mandatory = $true)]  [System.Boolean] $AuditingEnabled,
         [parameter(Mandatory = $false)] [System.UInt32]  $AuditlogMaxSize,